Split performance tests
